Restructure repo into frontend, backend, and old directories
