+++ old/streamlit_visual_editor_fixed.py
@@ -18,11 +18,11 @@ import requests
 load_dotenv()
 
 # Add current directory to path
-sys.path.append(os.path.dirname(os.path.abspath(__file__)))
+sys.path.append(os.path.join(os.path.dirname(os.path.abspath(__file__)), '..'))
 
 # Import components
-from main import create_wordpress_memory_agent
-from wordpress_tools import WordPressAPI
+from backend.main import create_wordpress_memory_agent
+from backend.wordpress_tools import WordPressAPI
 from langchain_core.messages import HumanMessage, AIMessage
 
 # Page config - MUST BE FIRST
